+++ runners/video-compressor.c
@@ -54,7 +54,7 @@
 #include <libavutil/rational.h>
 
 #define PROGRESS_PREFIX "__FFMPEG_WASM_PROGRESS__"
-#define RUNNER_VERSION "1.3.0"
+#define RUNNER_VERSION "1.4.0"
 
 typedef struct RunnerOptions {
     const char *input_path;