+++ runners/video-to-gif.c
@@ -1,0 +1,2 @@
+#define ANIMATION_FORMAT_GIF 1
+#include "video-to-animation-common.inc"