+++ runners/video-to-webp.c
@@ -1,0 +1,2 @@
+#define ANIMATION_FORMAT_WEBP 1
+#include "video-to-animation-common.inc"